+++ ext/src/TimestampGenerator/MonotonicGenerator/MonotonicGenerator.stub.php
@@ -1,0 +1,9 @@
+<?php
+
+/** @generate-class-entries */
+namespace Cassandra\TimestampGenerators {
+
+    final class MonotonicGenerator implements TimestampGeneratorInterface
+    {
+    }
+}