+++ ext/src/TimestampGenerator/TimestampGeneratorInterface/TimestampGeneratorInterface_arginfo.h
@@ -1,0 +1,19 @@
+/* This is a generated file, edit the .stub.php file instead.
+ * Stub hash: 04b95a632b093db14008352be1c386089c742a56 */
+
+
+
+
+static const zend_function_entry class_Cassandra_TimestampGenerators_TimestampGeneratorInterface_methods[] = {
+	ZEND_FE_END
+};
+
+static zend_class_entry *register_class_Cassandra_TimestampGenerators_TimestampGeneratorInterface(void)
+{
+	zend_class_entry ce, *class_entry;
+
+	INIT_NS_CLASS_ENTRY(ce, "Cassandra\\TimestampGenerators", "TimestampGeneratorInterface", class_Cassandra_TimestampGenerators_TimestampGeneratorInterface_methods);
+	class_entry = zend_register_internal_interface(&ce);
+
+	return class_entry;
+}