+++ ext/src/TimestampGenerator/TimestampGenerators.c
@@ -1,0 +1,14 @@
+#include <TimestampGenerator/TimestampGenerator.h>
+
+#include "MonotonicGenerator/MonotonicGenerator.h"
+#include "ServerSideGenerator/ServerSideGenerator.h"
+#include "TimestampGeneratorInterface/TimestampGeneratorInterface.h"
+
+void
+PhpDriverDefineTimestampGenerators()
+{
+  PhpDriverDefineTimestampGeneratorInterface();
+
+  PhpDriverDefineServerSideTimestampGenerator(phpDriverTimestampGeneratorInterfaceCe);
+  PhpDriverDefineMonotonicTimestampGenerator(phpDriverTimestampGeneratorInterfaceCe);
+}